refactor: move Rust tests into child modules
